refactor: factored out bindApplyTransformations, removed circular dependency
