+++ hrm-domain/hrm-core/case/caseService.ts
@@ -38,7 +38,7 @@ import { randomUUID } from 'crypto';
 import type { Contact } from '../contact/contactDataAccess';
 import { InitializedCan } from '../permissions/initializeCanForRules';
 import type { TwilioUser } from '@tech-matters/twilio-worker-auth';
-import { bindApplyTransformations as bindApplyContactTransformations } from '../contact/contactService';
+import { bindApplyTransformations as bindApplyContactTransformations } from '../contact/bindApplyTransformations';
 import type { Profile } from '../profile/profileDataAccess';
 import type { PaginationQuery } from '../search';
 import { TResult, newErr, newOk } from '@tech-matters/types';
