+++ hrm-domain/hrm-core/contact/bindApplyTransformations.ts
@@ -1,0 +1,54 @@
+/**
+ * Copyright (C) 2021-2023 Technology Matters
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see https://www.gnu.org/licenses/.
+ */
+
+import { isS3StoredTranscript } from '../conversation-media/conversation-media';
+import { actionsMaps } from '../permissions';
+import type { TwilioUser } from '@tech-matters/twilio-worker-auth';
+import type { InitializedCan } from '../permissions/initializeCanForRules';
+import type { Contact } from './contactDataAccess';
+
+const filterExternalTranscripts = (contact: Contact): Contact => {
+  const { conversationMedia, ...rest } = contact;
+  const filteredConversationMedia = (conversationMedia ?? []).filter(
+    m => !isS3StoredTranscript(m),
+  );
+
+  return {
+    ...rest,
+    conversationMedia: filteredConversationMedia,
+  };
+};
+
+type PermissionsBasedTransformation = {
+  action: (typeof actionsMaps)['contact'][keyof (typeof actionsMaps)['contact']];
+  transformation: (contact: Contact) => Contact;
+};
+
+const permissionsBasedTransformations: PermissionsBasedTransformation[] = [
+  {
+    action: actionsMaps.contact.VIEW_EXTERNAL_TRANSCRIPT,
+    transformation: filterExternalTranscripts,
+  },
+];
+
+export const bindApplyTransformations =
+  (can: InitializedCan, user: TwilioUser) =>
+  (contact: Contact): Contact =>
+    permissionsBasedTransformations.reduce(
+      (transformed, { action, transformation }) =>
+        !can(user, action, contact) ? transformation(transformed) : transformed,
+      contact,
+    );
